GL65_11/07_taxi request fix
done till driver can view and passenger also can view on the map for available drivers.
Need to adjust the variables which need to carry over the taxi request dto such as distance that need to be generated from google maps API on the place where the destination is selected.
best approach is in the time of taking the route, take it out from the map view so and derive other information needs from the response and store them and use them latter.